Added unit tests for the validation rules:
- Extracted base test fixture
